requirements.txt
Django==6.0.3
djangorestframework==3.14.0
